pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Ushbu loyiha BMI uchun backend."
authors = [
    {name = "Muhammadali", email = "mamadjonovmuhammadali7@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "aiogram (>=3.18.0,<4.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "flake8 (>=7.1.2,<8.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "tortoise-orm (>=0.25.0,<0.26.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "aerich (>=0.8.2,<0.9.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py311"]
exclude = '''
/(
    \.git
  | \.venv
  | venv
  | build
  | dist
  | migrations
)/
'''
